Grouping Toplevel events in ToplevelEventArgs.
